@@ -18,27 +18,28 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.1.0",
-    "@angular/cdk": "^13.0.0",
+    "@angular/animations": "^13.1.1",
+    "@angular/cdk": "^13.1.1",
     "@angular/common": "^13.1.1",
-    "@angular/compiler": "^13.0.0",
-    "@angular/core": "^13.0.1",
+    "@angular/compiler": "^13.1.1",
+    "@angular/core": "^13.1.1",
     "@angular/forms": "^13.1.1",
     "@angular/material": "^13.1.1",
     "@angular/platform-browser": "^13.1.1",
-    "@angular/platform-browser-dynamic": "^13.0.2",
+    "@angular/platform-browser-dynamic": "^13.1.1",
     "@angular/router": "^13.1.1",
-    "@splitsoftware/splitio-browserjs": "^0.2.0",
+    "@splitsoftware/splitio-browserjs": "^0.3.0",
+    "angular-gridster2": "^13.1.1",
     "ng-click-outside": "^9.0.1",
     "ngx-cookie-service": "^13.1.2",
-    "ngx-mask": "^13.0.1",
+    "ngx-mask": "^13.1.1",
     "rxjs": "^7.4.0",
     "tslib": "^2.3.1",
     "uuid": "^8.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1301.1",
+    "@angular-devkit/architect": "^0.1301.2",
     "@angular-devkit/build-angular": "^13.1.2",
     "@angular-eslint/builder": "^13.0.1",
     "@angular-eslint/eslint-plugin": "^13.0.1",
@@ -47,26 +48,26 @@
     "@angular-eslint/template-parser": "^13.0.1",
     "@angular/cli": "^13.1.2",
     "@angular/compiler-cli": "^13.1.1",
-    "@compodoc/compodoc": "^1.1.16",
+    "@compodoc/compodoc": "^1.1.18",
     "@strut-software/eslint-config": "^1.8.0",
     "@strut-software/stylelint-config": "^1.1.1",
     "@types/jasmine": "^3.10.3",
     "@types/node": "^16.11.19",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "5.9.0",
-    "@typescript-eslint/parser": "5.9.0",
+    "@typescript-eslint/eslint-plugin": "5.9.1",
+    "@typescript-eslint/parser": "5.9.1",
     "autoprefixer": "^10.4.2",
     "cross-env": "^7.0.3",
     "eslint": "8.6.0",
     "husky": "^7.0.4",
     "jasmine-core": "~4.0.0",
     "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.3.9",
+    "karma": "~6.3.10",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-mocks": "^12.5.1",
+    "ng-mocks": "^v13.0.0-alpha.1",
     "pa11y-ci": "3.0.1",
     "postcss": "^8.4.5",
     "postcss-cli": "^9.1.0",
@@ -75,7 +76,7 @@
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
     "stylelint-scss": "^3.21.0",
-    "tailwindcss": "^3.0.11",
+    "tailwindcss": "^3.0.13",
     "ts-node": "~10.4.0",
     "typescript": "~4.4.4"
   }
