@@ -1,9 +1,12 @@
-import { Component, Inject } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { BreakpointObserver } from '@angular/cdk/layout';
 import { StepperOrientation } from '@angular/material/stepper';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { first, map } from 'rxjs/operators';
+import { StationService } from 'src/app/core/station.service';
+import { ErrorService } from 'src/app/core/error.service';
+import { FlowLogicRule, Question } from 'src/models';
 /**
  * Reusable component for displaying the information to add a new rule.
  */
@@ -12,17 +15,31 @@ import { map } from 'rxjs/operators';
   templateUrl: './rule-modal.component.html',
   styleUrls: ['./rule-modal.component.scss'],
 })
-export class RuleModalComponent {
+export class RuleModalComponent implements OnInit {
   /** Station Rithm id. */
   stationRithmId = '';
 
   /** Orientation for stepper. */
   stepperOrientation$: Observable<StepperOrientation>;
 
+  /** The station Flow Logic Rule. */
+  stationFlowLogic!: FlowLogicRule;
+
+  /** The value of the first operand. */
+  firstOperand = '';
+
+  /** Get current and previous Questions for Stations. */
+  questionStation: Question[] = [];
+
+  /** The value of the operator. */
+  operator = '';
+
   constructor(
     public dialogRef: MatDialogRef<RuleModalComponent>,
     @Inject(MAT_DIALOG_DATA) public rithmId: string,
-    breakpointObserver: BreakpointObserver
+    breakpointObserver: BreakpointObserver,
+    private stationService: StationService,
+    private errorService: ErrorService
   ) {
     this.stationRithmId = rithmId;
     this.stepperOrientation$ = breakpointObserver
@@ -31,9 +48,56 @@ export class RuleModalComponent {
   }
 
   /**
+   * Life cycle init the component.
+   */
+  ngOnInit(): void {
+    this.getStationFlowLogicRule();
+  }
+
+  /**
    * Close rule Modal.
    */
   closeModal(): void {
     this.dialogRef.close();
+  }
+
+  /**
+   * Get each station flow rules.
+   */
+  private getStationFlowLogicRule(): void {
+    this.stationService
+      .getStationFlowLogicRule(this.stationRithmId)
+      .pipe(first())
+      .subscribe({
+        next: (stationFlowLogic) => {
+          this.stationFlowLogic = stationFlowLogic;
+        },
+        error: (error: unknown) => {
+          this.errorService.displayError(
+            "Something went wrong on our end and we're looking into it. Please try again in a little while.",
+            error
+          );
+        },
+      });
+  }
+
+  /**
+   * Get current and previous questions.
+   */
+  getStationQuestions(): void {
+    this.stationService
+      .getStationQuestions(this.stationRithmId, true)
+      .pipe(first())
+      .subscribe({
+        next: (questions) => {
+          this.questionStation = questions;
+        },
+        error: (error: unknown) => {
+          this.errorService.displayError(
+            "Something went wrong on our end and we're looking into it. Please try again in a little while.",
+            error
+          );
+        },
+      });
   }
 }
