@@ -1,5 +1,5 @@
 import { Component, Inject, OnInit } from '@angular/core';
-import { MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import {
   ConnectedModalData,
   ConnectedStationInfo,
@@ -8,6 +8,8 @@ import {
 import { DocumentService } from 'src/app/core/document.service';
 import { first } from 'rxjs';
 import { ErrorService } from 'src/app/core/error.service';
+import { PopupService } from 'src/app/core/popup.service';
+import { Router } from '@angular/router';
 
 /**
  * Component for connected stations.
@@ -36,10 +38,23 @@ export class ConnectedStationsModalComponent implements OnInit {
   /** The selected Station for move document. */
   selectedStation = '';
 
+  /* Load if exists error in the stations */
+  connectedError = false;
+
+  /* Loading in modal the list of connected stations */
+  connectedStationLoading = false;
+
+  /** Enable error message if move document request fails. */
+  moveDocumentError = false;
+
   constructor(
     @Inject(MAT_DIALOG_DATA) private data: ConnectedModalData,
     private documentService: DocumentService,
-    private errorService: ErrorService
+    private errorService: ErrorService,
+    private popupService: PopupService,
+    private matDialogRef: MatDialogRef<void>,
+    private router: Router,
+    public dialogRef: MatDialogRef<ConnectedStationsModalComponent>
   ) {
     this.documentRithmId = data.documentRithmId;
     this.stationRithmId = data.stationRithmId;
@@ -56,16 +71,20 @@ export class ConnectedStationsModalComponent implements OnInit {
    * Retrieves a list of the connected stations for the given document.
    */
   private getConnectedStations(): void {
+    this.connectedStationLoading = true;
     this.documentService
       .getConnectedStationInfo(this.documentRithmId, this.stationRithmId)
       .pipe(first())
       .subscribe({
         next: (connectedStations) => {
+          this.connectedStationLoading = false;
           this.connectedStations = connectedStations.nextStations.concat(
             connectedStations.previousStations
           );
         },
         error: (error: unknown) => {
+          this.connectedError = true;
+          this.connectedStationLoading = false;
           this.errorService.displayError(
             'Failed to get connected stations for this document.',
             error,
@@ -79,6 +98,7 @@ export class ConnectedStationsModalComponent implements OnInit {
    * Move the document from a station to another.
    */
   moveDocument(): void {
+    this.moveDocumentError = false;
     const moveDocument: MoveDocument = {
       fromStationRithmId: this.stationRithmId,
       toStationRithmIds: [this.selectedStation],
@@ -88,7 +108,14 @@ export class ConnectedStationsModalComponent implements OnInit {
       .moveDocument(moveDocument)
       .pipe(first())
       .subscribe({
+        next: () => {
+          this.popupService.notify('The document has been moved successfully');
+          this.moveDocumentError = false;
+          this.matDialogRef.close();
+          this.router.navigateByUrl('dashboard');
+        },
         error: (error: unknown) => {
+          this.moveDocumentError = true;
           this.errorService.displayError(
             "Something went wrong on our end and we're looking into it. Please try again in a little while.",
             error
