@@ -1,4 +1,9 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {
+  ComponentFixture,
+  fakeAsync,
+  TestBed,
+  tick,
+} from '@angular/core/testing';
 import { StationInfoDrawerComponent } from './station-info-drawer.component';
 import { StationService } from 'src/app/core/station.service';
 import {
@@ -21,7 +26,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatRadioModule } from '@angular/material/radio';
 import { LoadingIndicatorComponent } from 'src/app/shared/loading-indicator/loading-indicator.component';
 import { PopupService } from 'src/app/core/popup.service';
-import { DocumentGenerationStatus } from 'src/models';
+import { DialogOptions, DocumentGenerationStatus } from 'src/models';
 import { MapService } from 'src/app/map/map.service';
 import { DocumentService } from 'src/app/core/document.service';
 import { throwError } from 'rxjs';
@@ -156,8 +161,10 @@ describe('StationInfoDrawerComponent', () => {
   it('should show loading-indicators while get data the status station document', () => {
     component.stationLoading = false;
     component.getStationDocumentGenerationStatus();
+    spyOn(TestBed.inject(UserService), 'isStationOwner').and.returnValue(true);
     fixture.detectChanges();
     expect(component.docGenLoading).toBe(true);
+
     const loadingComponent = fixture.debugElement.nativeElement.querySelector(
       '#loading-indicator-status'
     );
@@ -171,6 +178,7 @@ describe('StationInfoDrawerComponent', () => {
       component.stationRithmId,
       newStatus
     );
+    spyOn(TestBed.inject(UserService), 'isStationOwner').and.returnValue(true);
     fixture.detectChanges();
     expect(component.docGenLoading).toBe(true);
     const loadingComponent = fixture.debugElement.nativeElement.querySelector(
@@ -214,18 +222,41 @@ describe('StationInfoDrawerComponent', () => {
     expect(valueExpected).toBeTrue();
   });
 
-  it('should create a document from station-info-drawer', () => {
+  it('should create a document', fakeAsync(() => {
+    const documentId = '78DF8E53-549E-44CD-8056-A2CBA055F32F';
     const createDocumentSpy = spyOn(
       TestBed.inject(DocumentService),
       'createNewDocument'
     ).and.callThrough();
+
+    const notifySpy = spyOn(
+      TestBed.inject(PopupService),
+      'notify'
+    ).and.callThrough();
+
+    const userService = TestBed.inject(UserService);
+    const user = userService.user;
+
+    const assignUser = spyOn(
+      component,
+      'assignUserToDocument'
+    ).and.callThrough();
+
     component.createNewDocument();
+    tick();
     expect(createDocumentSpy).toHaveBeenCalledOnceWith(
+      '',
+      0,
       component.stationRithmId
     );
-  });
+    tick(1000);
+    expect(notifySpy).toHaveBeenCalledOnceWith(
+      'The document has been created successfully.'
+    );
+    expect(assignUser).toHaveBeenCalledWith(user.rithmId, documentId);
+  }));
 
-  it('should catch an error if creating the document fails', () => {
+  it('should catch an error if creating the document fails', async () => {
     spyOn(TestBed.inject(DocumentService), 'createNewDocument').and.returnValue(
       throwError(() => {
         throw new Error();
@@ -236,11 +267,11 @@ describe('StationInfoDrawerComponent', () => {
       TestBed.inject(ErrorService),
       'displayError'
     ).and.callThrough();
-    component.createNewDocument();
+    await component.createNewDocument();
     expect(spyError).toHaveBeenCalled();
   });
 
-  it('should create new document and called services', () => {
+  it('should call the service to assign a user to a document', () => {
     const userExpect = '123-957';
     const newDocumentExpect = '852-789-654-782';
     const spyPetition = spyOn(
@@ -255,7 +286,7 @@ describe('StationInfoDrawerComponent', () => {
     );
   });
 
-  it('should catch error and executed error service', () => {
+  it('should catch error and executed error service when assign a user.', () => {
     const userExpect = '123-957';
     const newDocumentExpect = '852-789-654-782';
     spyOn(
@@ -272,5 +303,49 @@ describe('StationInfoDrawerComponent', () => {
     ).and.callThrough();
     component.assignUserToDocument(userExpect, newDocumentExpect);
     expect(spyError).toHaveBeenCalled();
+  });
+
+  it('should call the method createNewDocument when new-document button is clicked', fakeAsync(() => {
+    component.stationLoading = false;
+    component.stationDocumentGenerationStatus = DocumentGenerationStatus.Manual;
+
+    fixture.detectChanges();
+
+    const createDocumentSpy = spyOn(component, 'createNewDocument');
+    const btnNewDoc =
+      fixture.debugElement.nativeElement.querySelector('#new-document');
+    expect(btnNewDoc).toBeTruthy();
+    btnNewDoc.click();
+    tick();
+    expect(createDocumentSpy).toHaveBeenCalledOnceWith();
+  }));
+
+  it('should open a confirm dialog to create a document', async () => {
+    const dialogExpectData: DialogOptions = {
+      title: 'Are you sure?',
+      message:
+        'After the document is created you will be redirected to the document page.',
+      okButtonText: 'Confirm',
+      cancelButtonText: 'Cancel',
+    };
+    const popupSpy = spyOn(
+      TestBed.inject(PopupService),
+      'confirm'
+    ).and.callThrough();
+
+    await component.createNewDocument();
+    expect(popupSpy).toHaveBeenCalledOnceWith(dialogExpectData);
+  });
+
+  it('should show loading-indicators while creating a new document is underway', async () => {
+    component.stationLoading = false;
+    component.stationDocumentGenerationStatus = DocumentGenerationStatus.Manual;
+    await component.createNewDocument();
+    fixture.detectChanges();
+    expect(component.docCreationLoading).toBe(true);
+    const loadingComponent = fixture.debugElement.nativeElement.querySelector(
+      '#loading-indicator-status'
+    );
+    expect(loadingComponent).toBeTruthy();
   });
 });
