@@ -5,6 +5,11 @@ import { ConnectedStationInfo } from 'src/models';
 import { RuleModalComponent } from '../rule-modal/rule-modal.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatStepperModule } from '@angular/material/stepper';
+import { StationService } from 'src/app/core/station.service';
+import { MockErrorService, MockStationService } from 'src/mocks';
+import { ErrorService } from 'src/app/core/error.service';
+import { MatSelectModule } from '@angular/material/select';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 describe('FlowLogicComponent', () => {
   let component: FlowLogicComponent;
@@ -19,8 +24,18 @@ describe('FlowLogicComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [MatDialogModule, BrowserAnimationsModule, MatStepperModule],
+      imports: [
+        MatDialogModule,
+        BrowserAnimationsModule,
+        MatStepperModule,
+        MatSelectModule,
+        MatSnackBarModule,
+      ],
       declarations: [FlowLogicComponent, RuleModalComponent],
+      providers: [
+        { provide: StationService, useClass: MockStationService },
+        { provide: ErrorService, useClass: MockErrorService },
+      ],
     }).compileComponents();
   });
 
