@@ -20,7 +20,7 @@ import { MatCheckboxChange } from '@angular/material/checkbox';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { StationService } from 'src/app/core/station.service';
-import { Question, QuestionFieldType } from 'src/models';
+import { Question, QuestionFieldType, PossibleAnswer } from 'src/models';
 
 /**
  * Station Field Component.
@@ -89,22 +89,10 @@ export class StationFieldComponent
     value: '',
   };
 
-  /** The field for adding an option to a select field. */
-  selectOptionField: Question = {
+  /** The field for adding an option to a selectable fieldQuestionType. */
+  selectableOption: Question = {
     rithmId: '',
     prompt: 'Add Option',
-    questionType: QuestionFieldType.ShortText,
-    isReadOnly: false,
-    isRequired: true,
-    isPrivate: false,
-    children: [],
-    isPossibleAnswer: true,
-  };
-
-  /** The field for adding an item to a checklist field. */
-  checklistOptionField: Question = {
-    rithmId: '',
-    prompt: 'Add Item',
     questionType: QuestionFieldType.ShortText,
     isReadOnly: false,
     isRequired: true,
@@ -141,15 +129,19 @@ export class StationFieldComponent
       this.field.questionType === this.fieldType.MultiSelect ||
       this.field.questionType === this.fieldType.CheckList
     ) {
-      this.addOption(this.field.questionType);
+      if (this.field.possibleAnswers && this.field.possibleAnswers.length) {
+        this.populateAnswers(this.field.possibleAnswers);
+      } else {
+        this.addOption(this.field.questionType);
+      }
     }
     this.stationFieldForm = this.fb.group({
       instructionsField: [''],
       [this.field.questionType]: [''],
       optionField: [''],
-      isRequired: [this.field.isRequired],
-      isPrivate: [this.field.isPrivate],
-      isReadOnly: [this.field.isReadOnly],
+      [`isRequired-${this.field.rithmId}`]: [this.field.isRequired],
+      [`isPrivate-${this.field.rithmId}`]: [this.field.isPrivate],
+      [`isReadonly-${this.field.rithmId}`]: [this.field.isReadOnly],
     });
     this.stationFieldForm.valueChanges
       .pipe(takeUntil(this.destroyed$))
@@ -192,6 +184,35 @@ export class StationFieldComponent
   }
 
   /**
+   * Display existing possibles Answers.
+   *
+   * @param answers THe array of answers to be shown.
+   */
+  private populateAnswers(answers: PossibleAnswer[]): void {
+    answers.forEach((answer) => {
+      const possibleAnswer: Question = {
+        rithmId: this.field.rithmId, //Rithm id of the question to track the answer
+        prompt:
+          this.field.questionType === QuestionFieldType.MultiSelect ||
+          this.field.questionType === QuestionFieldType.Select
+            ? 'Add option'
+            : this.field.questionType === QuestionFieldType.CheckList
+            ? 'Add item'
+            : '',
+        questionType: QuestionFieldType.ShortText,
+        isReadOnly: false,
+        isRequired: true,
+        isPrivate: false,
+        children: [],
+        isPossibleAnswer: true,
+        value: answer.text,
+        originalStationRithmId: answer.rithmId, //rithm id of the answer to update the answer
+      };
+      this.options.push(possibleAnswer);
+    });
+  }
+
+  /**
    * Add an option to the options array.
    *
    * @param fieldType The field type.
@@ -201,16 +222,19 @@ export class StationFieldComponent
       [...Array(size)]
         .map(() => Math.floor(Math.random() * 16).toString(16))
         .join('');
-    const answerRithmId = `ans-${genRanHex(6)}`;
-    this.selectOptionField.rithmId = this.field.rithmId;
-    this.checklistOptionField.rithmId = this.field.rithmId;
-    this.selectOptionField.originalStationRithmId = answerRithmId;
-    this.checklistOptionField.originalStationRithmId = answerRithmId;
-    this.options.push(
+    const answerRithmId = `ans-${genRanHex(8)}-${genRanHex(4)}-${genRanHex(
+      4
+    )}-${genRanHex(4)}-${genRanHex(12)}`;
+    this.selectableOption.prompt =
+      fieldType === QuestionFieldType.MultiSelect ||
       fieldType === QuestionFieldType.Select
-        ? this.selectOptionField
-        : this.checklistOptionField
-    );
+        ? 'Add option'
+        : fieldType === QuestionFieldType.CheckList
+        ? 'Add item'
+        : '';
+    this.selectableOption.rithmId = this.field.rithmId;
+    this.selectableOption.originalStationRithmId = answerRithmId;
+    this.options.push(this.selectableOption);
   }
 
   /**
@@ -223,6 +247,7 @@ export class StationFieldComponent
     if (this.field.possibleAnswers) {
       this.field.possibleAnswers.splice(index, 1);
     }
+    this.stationService.touchStationForm();
   }
 
   /**
