@@ -1,7 +1,7 @@
 /**
- * Represents all the types of hover interactions on a flow in the map.
+ * Represents all the items of hover interactions on a flow in the map.
  */
-export enum FlowElementHoverType {
+export enum FlowElementHoverItem {
   /** The map cursor is hovering over the flow boundary. */
   Boundary = 'boundary',
 
