@@ -2,11 +2,15 @@ export * from './email-link-type.enum';
 export * from './user-type.enum';
 export * from './question-field-type.enum';
 export * from './map-mode.enum';
-export * from './station-element-hover-type.enum';
+export * from './station-element-hover-item.enum';
 export * from './map-item-status.enum';
 export * from './map-drag-item.enum';
 export * from './dialog-type.enum';
 export * from './document-generation-status.enum';
-export * from './flow-element-hover-type.enum';
+export * from './flow-element-hover-item.enum';
 export * from './field-name-separator.enum';
 export * from './environment-name.enum';
+export * from './widget-type.enum';
+export * from './operand-type.enum';
+export * from './operator-type.enum';
+export * from './rule-type.enum';
